requirements.txt
asn1crypto==0.24.0
astroid==2.2.5
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
cryptography==2.7
Flask==1.1.1
future==0.17.1
idna==2.8
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.1
MarkupSafe==1.1.1
mccabe==0.6.1
pycparser==2.19
pylint==2.3.1
python-telegram-bot==11.1.0
requests==2.22.0
six==1.12.0
typed-ast==1.4.0
urllib3==1.25.3
Werkzeug==0.15.5
wrapt==1.11.2
gunicorn
wikipedia
googletrans
unidecode
pyowm
opencv-python==3.4.1.15
matplotlib
numpy
Pillow